2001-07-10 01:55 UTC-0800 Ron Pinkas <ron@profit-master.com>    * source/compiler/cmdcheck.c      + Added missing type cast to resolve gcc warning.    * source/pp/ppcore.c      + Added missing type cast to resolve gcc warning.      - Removed 4 redundant #includes already included by included hbcomp.h

+++ harbour/source/compiler/cmdcheck.c
@@ -262,7 +262,7 @@ void hb_compChkCompilerSwitch( int iArg, char * Args[] )
                      case 'G' :
                           /* Required argument */
                           Switch[2] = Args[i][j + 1];
-                          if( isdigit( Args[i][j + 2] ) )
+                          if( isdigit( (int) Args[i][j + 2] ) )
                           {
                              /* Optional argument */
                              Switch[3] = Args[i][j + 2];
